requirements.txt
--index-url https://pypi.org/simple
--extra-index-url https://www.paddlepaddle.org.cn/packages/stable/cu129/

# Paddle Core
paddlepaddle-gpu==3.2.0
paddleocr==3.3.0
paddlex==3.3.4

# Compatible dependencies for Python 3.10
numpy<2.0,>=1.26.0
networkx<3.5,>=3.4.0
pillow>=10.0.0
opencv-contrib-python==4.10.0.84
protobuf>=4,<7
ujson>=5.10.0
PyYAML>=6
tqdm>=4.64
prettytable>=3.0

# Flask 백엔드 의존성
Flask==3.0.0
flask-cors==4.0.0
Werkzeug==3.0.1

# API 요청
requests==2.31.0
python-dotenv==1.0.0